+++ Modules/Calcs/ppd_prod.py
@@ -14,7 +14,7 @@ def fert__multiplier_array():
         else:
             fert = fert*(10/33) + 1
         fert_list.append(fert)
-    return np.array(fert_list)
+    return np.array(fert_list, dtype=np.float32)
 
 
 def bui_multiple_array():
@@ -23,4 +23,4 @@ def bui_multiple_array():
     for prod_bui in default_prod_buildings:
         mult = planner[prod_bui][prod_bui]
         bui_list.append(mult)
-    return np.array(bui_list)
+    return np.array(bui_list, dtype=np.float32)
