+++ Modules/FIO/workforce_req.py
@@ -21,6 +21,6 @@ def workforce_req():
             amount = consum['Amount']
             temp_list[index_pos] = amount
             
-        req_dict[worker_name] = np.array(temp_list)
+        req_dict[worker_name] = np.array(temp_list, dtype=np.float32)
 
     return req_dict
